Restructure PDF generation and refactor device categorization
Move pdf.go to pdf/pdf.go and update package to pdf for better organization. Introduce new device categories - ineligibleHardware, unsupportedVersions, and registrationCandidates - replacing affectedDevices and unaffectedDevices. Add helper function addDevicesTable to streamline table creation and improve code readability.
